requirements.txt
aiohttp==3.8.1
aiohttp-jinja2==1.5
aiosignal==1.2.0
async-timeout==4.0.2
asyncpg==0.25.0
attrs==21.4.0
charset-normalizer==2.0.12
frozenlist==1.3.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.8
pytest==7.1.1
pytest-aiohttp==1.0.4
pytest-asyncio==0.18.3
PyYAML==6.0
tomli==2.0.1
yarl==1.7.2
